backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
azure-core==1.34.0
azure-identity==1.23.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
google-api-core==2.25.1
google-auth==2.23.4
google-cloud-core==2.4.3
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-cloud-texttospeech==2.27.0
google-cloud-translate==3.12.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
motor==3.3.2
msal==1.32.3
msal-extensions==1.3.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.10
opencensus-ext-logging==0.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
proto-plus==1.26.1
protobuf==4.25.8
psutil==5.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.10.1
pymongo==4.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
requests==2.32.4
rsa==4.9.1
sentry-sdk==1.38.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==12.0
